fix(review): 提取 _resolve_file_type 到独立模块，测试直接导入生产代码
- 新增 api/api/_file_type.py（无模块级副作用，可安全导入）
- node_graph.py 改为 from _file_type import _resolve_file_type
- 测试移至 tests/api/（项目惯例），直接导入 _resolve_file_type
- 删除 tests/unit/api/test_node_graph_file_type.py（复制逻辑的旧测试）

Review: https://github.com/Kaoruha/GinkGO/pull/6025#issuecomment-4673380117
